Conflicts:
	.gitignore
	src/main/java/com/sequoiagrove/controller/ScheduleController.java

+++ src/main/webapp/WEB-INF/js/controllers/schedule.js
@@ -20,21 +20,23 @@ angular.module('sequoiaGroveApp')
         $translate) {
 
 /************** Login Redirect, Containers and UI settings **************/
+  $rootScope.lastPath = '/schedule';
+  $scope.saving = false;
+  $scope.importing = false;
 
   // user is not logged in
   if ($rootScope.loggedIn == false) {
     $location.path('/login');
   }
-  $rootScope.lastPath = '/schedule';
 
   $scope.activeTab = 'schedule';
   $scope.selectedId = 0;
   $scope.newDelivery = '';
   $scope.selectedPid = 0;
   $scope.empEditSearch = '';
-  $scope.saving = false;
   $scope.selectedShift = {
     idx : -1,
+    sid : -1,
     title : '',
     pos : '',
     wd_st : '',
@@ -57,16 +59,27 @@ angular.module('sequoiaGroveApp')
   $scope.selectShift = function(cur) {
     $scope.selectedShift.idx = cur;
     if ($scope.selectedShift.idx != -1) {
-      /*set title*/
+      $scope.selectedShift.sid = $scope.template[cur].sid;
       $scope.selectedShift.title = $scope.template[cur].tname;
-      /*set position*/
       $scope.selectedShift.pos = $scope.template[cur].position;
-      /*set hours*/
       $scope.selectedShift.wd_st = moment($scope.template[cur].wd_st_h + ':' + $scope.template[cur].wd_st_m, 'HH:mm').format('h:mm A');
       $scope.selectedShift.wd_ed = moment($scope.template[cur].wd_ed_h + ':' + $scope.template[cur].wd_ed_m, 'HH:mm').format('h:mm A');
       $scope.selectedShift.we_st = moment($scope.template[cur].we_st_h + ':' + $scope.template[cur].we_st_m, 'HH:mm').format('h:mm A');
       $scope.selectedShift.we_ed = moment($scope.template[cur].we_ed_h + ':' + $scope.template[cur].we_ed_m, 'HH:mm').format('h:mm A');
     }
+    else {
+      $scope.selectedShift.sid = 0;
+      $scope.selectedShift.title = '';
+      $scope.selectedShift.pos = '';
+      $scope.selectedShift.wd_st = '';
+      $scope.selectedShift.wd_ed = '';
+      $scope.selectedShift.we_st = '';
+      $scope.selectedShift.we_ed = '';
+    }
+  }
+
+  $scope.shiftSelected = function() {
+    return ($scope.selectedShift.idx != -1);
   }
 
   $scope.switchPos = function(pos) {
@@ -152,7 +165,7 @@ angular.module('sequoiaGroveApp')
 
     // 3. check employee availability against shift duration
     _.map(avail, function(a, index) {
-      if (a.start.isBefore(shiftStart) && a.end.isAfter(shiftEnd)) {
+      if ((a.start.isBefore(shiftStart) || a.start.isSame(shiftStart)) && (a.end.isAfter(shiftEnd) || a.end.isSame(shiftEnd))) {
         isAvailable = true;
       }
     });
@@ -298,7 +311,7 @@ angular.module('sequoiaGroveApp')
       $scope.updateShifts.push(paramObj);
     }
 
-    $scope.selectEid(eid);
+    $scope.selectedId = eid;
   }
 
 
@@ -343,31 +356,38 @@ angular.module('sequoiaGroveApp')
   }
 
   $scope.importLastWeek = function() {
+    $scope.importing = true;
+    $scope.selectedId = 0;
     var d = moment($scope.date.mon.val,'DD-MM-YYYY').subtract(7, 'days').format('DD-MM-YYYY');
-    $scope.getScheduleTemplate(d);
+   $scope.getScheduleTemplate(d)
+     .then(function() {
+       $scope.importing = false;
+   });
   }
 
 /************** HTTP Request Functions **************/
 
-  // Save the shifts in the list of updateShifts
+  // Save these shift schedulings in the list of updateShifts
   $scope.saveSchedule = function() {
+    if ($scope.saving) {
+      return;
+    }
+    $scope.saving = true;
+    $scope.selectedId = 0;
     // remove blank spaces from update list - they are in delete shifts, or
     // have not been assigned
     $scope.updateShifts = _.filter($scope.updateShifts, function(shift) {
       return (shift.eid !== 0);
     });
-    $log.debug($scope.updateShifts);
-    $scope.saving = true;
 
     $http({
       url: '/sequoiagrove/schedule/update/',
       method: "POST",
-      data: { 'body': JSON.stringify($scope.updateShifts) }
+      data: $scope.updateShifts
     }).success(function (data, status, headers, config) {
       if (status == 200) {
         // clear update shifts list
         $scope.updateShifts = [];
-        $scope.saving = false;
         $scope.deleteSchedule();
       }
       else {
@@ -379,14 +399,15 @@ angular.module('sequoiaGroveApp')
     });
   }
 
-  // Delete all these shifts
+  // Delete these shift schedulings
   $scope.deleteSchedule = function() {
-    $scope.saving = true;
-
+    if ($scope.saving) {
+      return;
+    }
     $http({
       url: '/sequoiagrove/schedule/delete/',
       method: "DELETE",
-      data: { 'body': JSON.stringify($scope.deleteShifts) }
+      data: $scope.deleteShifts
     }).success(function (data, status, headers, config) {
       if (status == 200) {
         // clear delete shifts list
@@ -402,6 +423,10 @@ angular.module('sequoiaGroveApp')
     });
   }
 
+  // Add new shift to schedule
+  $scope.addShift = function() {
+  }
+
 /************** Controller Initialization **************/
 
   $scope.init = function() {
@@ -411,9 +436,9 @@ angular.module('sequoiaGroveApp')
 
 /************** Event Watchers **************/
 
-  $scope.$watch('selectedId', function(newVal, oldVal){
+  $scope.$watch($rootScope.loading, function(newVal, oldVal){
     if(newVal){
-      $log.debug(newVal);
+      //$log.debug(newVal);
       // watchExpression has changed.
     }
   });
